refactor: update script.js and related modules to improve state management and streamline function parameters
