+++ eslint.config.js
@@ -1,0 +1,146 @@
+export default [
+  {
+    ignores: ['**/*.config.js'],
+    languageOptions: {
+      ecmaVersion: 2024,
+      sourceType: 'module',
+      globals: {
+        // Browser globals
+        window: 'readonly',
+        document: 'readonly',
+        navigator: 'readonly',
+        console: 'readonly',
+        localStorage: 'readonly',
+        sessionStorage: 'readonly',
+        fetch: 'readonly',
+        URL: 'readonly',
+        URLSearchParams: 'readonly',
+        FormData: 'readonly',
+        Headers: 'readonly',
+        Request: 'readonly',
+        Response: 'readonly',
+        Event: 'readonly',
+        CustomEvent: 'readonly',
+        EventTarget: 'readonly',
+        MutationObserver: 'readonly',
+        IntersectionObserver: 'readonly',
+        ResizeObserver: 'readonly',
+        AbortController: 'readonly',
+        AbortSignal: 'readonly',
+        Blob: 'readonly',
+        File: 'readonly',
+        FileReader: 'readonly',
+        WebSocket: 'readonly',
+        Worker: 'readonly',
+        ServiceWorker: 'readonly',
+        // PWA specific
+        caches: 'readonly',
+        indexedDB: 'readonly',
+        IDBDatabase: 'readonly',
+        IDBTransaction: 'readonly',
+        IDBObjectStore: 'readonly',
+        IDBRequest: 'readonly',
+        IDBKeyRange: 'readonly',
+        // Timer functions
+        setTimeout: 'readonly',
+        clearTimeout: 'readonly',
+        setInterval: 'readonly',
+        clearInterval: 'readonly',
+        requestAnimationFrame: 'readonly',
+        cancelAnimationFrame: 'readonly',
+        requestIdleCallback: 'readonly',
+        cancelIdleCallback: 'readonly',
+        // Browser dialogs
+        alert: 'readonly',
+        confirm: 'readonly',
+        prompt: 'readonly',
+        // Crypto API
+        crypto: 'readonly',
+        // Cypress global
+        cy: 'readonly',
+        Cypress: 'readonly',
+        // Cypress test globals
+        describe: 'readonly',
+        it: 'readonly',
+        beforeEach: 'readonly',
+        expect: 'readonly',
+        // Browser API
+        getBoundingClientRect: 'readonly',
+        getComputedStyle: 'readonly',
+        // Service worker
+        self: 'readonly'
+      }
+    },
+    rules: {
+      // Possible Problems
+      'no-unused-vars': ['warn', { 'argsIgnorePattern': '^_$|^_', 'varsIgnorePattern': '^_$|^_' }],
+      'no-undef': 'error',
+      'no-unreachable': 'error',
+      'no-constant-condition': 'error',
+      'no-duplicate-case': 'error',
+      'no-empty': ['error', { 'allowEmptyCatch': true }],
+      'no-extra-boolean-cast': 'error',
+      'no-extra-semi': 'error',
+      'no-func-assign': 'error',
+      'no-inner-declarations': 'error',
+      'no-invalid-regexp': 'error',
+      'no-irregular-whitespace': 'error',
+      'no-obj-calls': 'error',
+      'no-sparse-arrays': 'error',
+      'no-unexpected-multiline': 'error',
+      'use-isnan': 'error',
+      'valid-typeof': 'error',
+
+      // Best Practices
+      'eqeqeq': ['error', 'always'],
+      'no-eval': 'error',
+      'no-implied-eval': 'error',
+      'no-global-assign': 'error',
+      'no-implicit-globals': 'error',
+      'no-loop-func': 'error',
+      'no-new-func': 'error',
+      'no-new-wrappers': 'error',
+      'no-octal': 'error',
+      'no-redeclare': 'error',
+      'no-self-assign': 'error',
+      'no-self-compare': 'error',
+      'no-throw-literal': 'error',
+      'no-unused-expressions': 'error',
+      'no-useless-call': 'error',
+      'no-useless-concat': 'error',
+      'no-useless-escape': 'error',
+      'no-useless-return': 'error',
+      'no-with': 'error',
+      'radix': 'error',
+
+      // Stylistic Issues
+      'indent': ['error', 2],
+      'quotes': ['error', 'single', { 'avoidEscape': true }],
+      'semi': ['error', 'always'],
+      'comma-dangle': ['error', 'never'],
+      'comma-spacing': ['error', { 'before': false, 'after': true }],
+      'comma-style': ['error', 'last'],
+      'key-spacing': ['error', { 'beforeColon': false, 'afterColon': true }],
+      'no-trailing-spaces': 'error',
+      'no-multiple-empty-lines': ['error', { 'max': 2 }],
+      'space-before-blocks': 'error',
+      'space-in-parens': ['error', 'never'],
+      'space-infix-ops': 'error',
+      'keyword-spacing': 'error',
+      'object-curly-spacing': ['error', 'always'],
+      'array-bracket-spacing': ['error', 'never'],
+      'brace-style': ['error', '1tbs', { 'allowSingleLine': true }],
+
+      // ES6
+      'no-var': 'error',
+      'prefer-const': 'error',
+      'prefer-arrow-callback': 'error',
+      'arrow-spacing': 'error',
+      'no-duplicate-imports': 'error',
+      'no-useless-constructor': 'error',
+      'no-useless-rename': 'error',
+      'object-shorthand': 'error',
+      'prefer-template': 'error'
+    }
+  }
+];
